+++ ui/js/actions/content.js
@@ -4,7 +4,7 @@ import lbry from "lbry";
 import lbryio from "lbryio";
 import lbryuri from "lbryuri";
 import { makeSelectClientSetting } from "selectors/settings";
-import { selectBalance } from "selectors/wallet";
+import { selectBalance, selectTransactionItems } from "selectors/wallet";
 import {
   makeSelectFileInfoForUri,
   selectDownloadingByOutpoint,
@@ -493,5 +493,48 @@ export function doPublish(params) {
 
       lbry.publishDeprecated(params, null, success, failure);
     });
+  };
+}
+
+export function doAbandonClaim(txid, nout) {
+  return function(dispatch, getState) {
+    const state = getState();
+    const transactionItems = selectTransactionItems(state);
+    const { claim_id: claimId, claim_name: name } = transactionItems.find(
+      claim => claim.txid == txid && claim.nout == nout
+    );
+
+    dispatch({
+      type: types.ABANDON_CLAIM_STARTED,
+      data: {
+        claimId: claimId,
+      },
+    });
+
+    const errorCallback = error => {
+      dispatch(doOpenModal(modals.TRANSACTION_FAILED));
+    };
+
+    const successCallback = results => {
+      if (results.txid) {
+        dispatch({
+          type: types.ABANDON_CLAIM_SUCCEEDED,
+          data: {
+            claimId: claimId,
+          },
+        });
+        dispatch(doResolveUri(lbryuri.build({ name, claimId })));
+        dispatch(doFetchClaimListMine());
+      } else {
+        dispatch(doOpenModal(modals.TRANSACTION_FAILED));
+      }
+    };
+
+    lbry
+      .claim_abandon({
+        txid: txid,
+        nout: nout,
+      })
+      .then(successCallback, errorCallback);
   };
 }