+++ ui/js/component/transactionList/view.jsx
@@ -1,6 +1,9 @@
 import React from "react";
 import TransactionListItem from "./internal/TransactionListItem";
 import FormField from "component/formField";
+import Link from "component/link";
+import * as icons from "constants/icons";
+import * as modals from "constants/modal_types";
 
 class TransactionList extends React.PureComponent {
   constructor(props) {
@@ -21,6 +24,16 @@ class TransactionList extends React.PureComponent {
     const { filter } = this.state;
 
     return !filter || filter == transaction.type;
+  }
+
+  isRevokeable(txid, nout) {
+    // a claim/support/update is revokable if it
+    // is in my claim list(claim_list_mine)
+    return this.props.myClaims.has(`${txid}:${nout}`);
+  }
+
+  revokeClaim(txid, nout) {
+    this.props.openModal(modals.CONFIRM_CLAIM_REVOKE, { txid, nout });
   }
 
   render() {
@@ -48,6 +61,11 @@ class TransactionList extends React.PureComponent {
               <option value="support">{__("Supports")}</option>
               <option value="update">{__("Updates")}</option>
             </FormField>
+            {" "}
+            <Link
+              href="https://lbry.io/faq/transaction-types"
+              icon={icons.HELP_CIRCLE}
+            />
           </span>}
         {!transactionList.length &&
           <div className="empty">
@@ -70,6 +88,8 @@ class TransactionList extends React.PureComponent {
                   key={`${t.txid}:${t.nout}`}
                   transaction={t}
                   reward={rewards && rewards[t.txid]}
+                  isRevokeable={this.isRevokeable(t.txid, t.nout)}
+                  revokeClaim={this.revokeClaim.bind(this)}
                 />
               )}
             </tbody>