+++ ui/js/component/video/view.jsx
@@ -73,9 +73,7 @@ class Video extends React.PureComponent {
         "It looks like you deleted or moved this file. We're rebuilding it now. It will only take a few seconds."
       );
     } else if (isLoading) {
-      loadStatusMessage = __(
-        "Requesting stream..."
-      );
+      loadStatusMessage = __("Requesting stream...");
     } else if (isDownloading) {
       loadStatusMessage = __("Downloading stream... not long left now!");
     }