+++ ui/js/constants/icons.js
@@ -2,3 +2,4 @@ export const FEATURED = "rocket";
 export const LOCAL = "folder";
 export const FILE = "file";
 export const HISTORY = "history";
+export const HELP_CIRCLE = "question-circle";