+++ ui/js/modal/modalRevokeClaim/index.js
@@ -1,0 +1,17 @@
+import React from "react";
+import { connect } from "react-redux";
+import { doCloseModal } from "actions/app";
+import { doAbandonClaim } from "actions/content";
+import { selectTransactionItems } from "selectors/wallet";
+import ModalRevokeClaim from "./view";
+
+const select = state => ({
+  transactionItems: selectTransactionItems(state),
+});
+
+const perform = dispatch => ({
+  closeModal: () => dispatch(doCloseModal()),
+  abandonClaim: (txid, nout) => dispatch(doAbandonClaim(txid, nout)),
+});
+
+export default connect(select, perform)(ModalRevokeClaim);